requirements.txt
awsebcli==3.14.1
Babel==2.7.0
blessed==1.15.0
cached-property==1.4.3
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.9
django-cors-headers==3.0.2
django-heroku==0.3.1
django-phonenumber-field==3.0.1
django-twilio==0.10.0
djangorestframework==3.9.4
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
jmespath==0.9.3
pathspec==0.5.5
psycopg2==2.7.7
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
semantic-version==2.5.0
six==1.12.0
sqlparse==0.3.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
twilio==6.28.0
urllib3==1.25.3
websocket-client==0.48.0
whitenoise==4.1.2
